requirements.txt
asyncpg==0.20.1
click==7.1.2
databases==0.3.2
fastapi==0.58.1
gunicorn==20.0.4
h11==0.9.0
psycopg2-binary==2.8.5
pydantic==1.5.1
SQLAlchemy==1.3.18
starlette==0.13.4
uvicorn==0.11.5
websockets==8.1
